Add helpers.go for some common test utils.
